move closure back vm
