+++ include/tvm/runtime/vm.h
@@ -25,17 +25,51 @@
 #define TVM_RUNTIME_VM_H_
 
 #include <tvm/runtime/object.h>
-#include <tvm/runtime/common_object.h>
+#include <tvm/runtime/memory.h>
 #include <tvm/runtime/packed_func.h>
 #include <tvm/runtime/registry.h>
 #include <memory>
 #include <string>
 #include <unordered_map>
+#include <utility>
 #include <vector>
 
 namespace tvm {
 namespace runtime {
 namespace vm {
+
+/*!
+ * \brief An object representing a closure. This object is used by both the
+ * Relay VM and interpreter.
+ */
+class ClosureObj : public Object {
+ public:
+  /*!
+   * \brief The index into the function list. The function could be any
+   * function object that is compatible to a certain runtime, i.e. VM or
+   * interpreter.
+   */
+  size_t func_index;
+  /*! \brief The free variables of the closure. */
+  std::vector<ObjectRef> free_vars;
+
+  static constexpr const uint32_t _type_index = TypeIndex::kClosure;
+  static constexpr const char* _type_key = "Closure";
+  TVM_DECLARE_FINAL_OBJECT_INFO(ClosureObj, Object);
+};
+
+/*! \brief reference to closure. */
+class Closure : public ObjectRef {
+ public:
+  Closure(size_t func_index, std::vector<ObjectRef> free_vars) {
+    auto ptr = make_object<ClosureObj>();
+    ptr->func_index = func_index;
+    ptr->free_vars = std::move(free_vars);
+    data_ = std::move(ptr);
+  }
+
+  TVM_DEFINE_OBJECT_REF_METHODS(Closure, ObjectRef, ClosureObj);
+};
 
 /*! \brief Magic number for NDArray list file  */
 constexpr uint64_t kTVMNDArrayListMagic = 0xF7E58D4F05049CB7;
